refactor packages using standard notation
