+++ src/main/java/com/wireshield/virustotal/VirusTotalScanner.java
@@ -1,4 +1,4 @@
-package com.wireshield;
+package com.wireshield.virustotal;
 
 /**
  * Scansione antivirus tramite API di VirusTotal.
